+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Beaker.workspace.experiments()`.
 - Added `Beaker.workspace.images()`.
 
+### Removed
+
+- Removed `Beaker.experiment.list()`. Please use `Beaker.workspace.experiments()` instead.
+
 ## [v0.5.5](https://github.com/allenai/beaker-py/releases/tag/v0.5.5) - 2022-04-10
 
 ### Added